second

+++ PriceState.Interfaces/Model/Region/GetRegionRequest.cs
@@ -4,7 +4,7 @@ namespace PriceState.Interfaces.Model.Region;
 
 public class GetRegionRequest: IPaginationRequest
 {
-    public long? RegionId { get; set; } = null;
+    public int? RegionId { get; set; } = null;
 
     public Page Page { get; set; } = new Page();
 }
